+++ src/main/java/com/ninjaone/backendinterviewproject/dto/input/RmmServiceExecutionRequestDto.java
@@ -2,26 +2,26 @@ package com.ninjaone.backendinterviewproject.dto.input;
 
 public class RmmServiceExecutionRequestDto {
 
-    private final Integer id = null;
-    private final Integer deviceId;
-    private final Integer rmmServiceId;
+    private final Long id = null;
+    private final Long deviceId;
+    private final Long rmmServiceId;
     private final Integer quantity;
 
-    public RmmServiceExecutionRequestDto(Integer deviceId, Integer rmmServiceId, Integer quantity) {
+    public RmmServiceExecutionRequestDto(Long deviceId, Long rmmServiceId, Integer quantity) {
         this.deviceId = deviceId;
         this.rmmServiceId = rmmServiceId;
         this.quantity = quantity;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public Integer getDeviceId() {
+    public Long getDeviceId() {
         return deviceId;
     }
 
-    public Integer getRmmServiceId() {
+    public Long getRmmServiceId() {
         return rmmServiceId;
     }
